mvoe methods from miner to base node && rename tx_startegy -> tx_modeling
