+++ sim/base_models.py
@@ -1,7 +1,7 @@
 """
 The classes that make up the core simulator.
 """
-
+import math
 from enum import Enum
 
 from loguru import logger
@@ -9,8 +9,7 @@ from loguru import logger
 from typing import List, Dict
 
 from sim import util
-from sim.network_util import speed, latency
-from sim.util import Region
+from sim.network_util import get_delay, Region
 
 
 class Item:
@@ -48,6 +47,7 @@ class Node:
         self.id = util.generate_uuid()
         self.timestamp = timestamp
         self.region = region
+        self.iter_seconds = 0.1
 
         self.inbox: Dict[int, List[Packet]] = dict()
         """Node's inbox with simulation timestamps as keys and lists of `Item`s to be consumed at that timestamp as values."""
@@ -85,3 +85,28 @@ class Node:
         self.ins = dict()
         self.outs = dict()
         self.last_reveal_times = dict()
+
+    def send_to(self, node, item: Item):
+        """
+        Send an item to a specific node. Can be used to respond to messages.
+        * node (`sim.base_models.Node`): Target node.
+        * item (`sim.base_models.Item`): Item to send.
+        """
+        packet = Packet(item)
+        delay = get_delay(self.region, node.region, item.size) / self.iter_seconds
+        reveal_time = math.ceil(max(self.timestamp, self.last_reveal_times.get(node.id, 0)) + delay)
+        self.last_reveal_times[node.id] = reveal_time
+        packet.reveal_at = reveal_time
+        try:
+            node.inbox[packet.reveal_at].append(packet)
+        except KeyError:
+            node.inbox[packet.reveal_at] = [packet]
+
+    def connect(self, *argv):
+        """
+        Establish an outgoing connection to one or more nodes.
+        * argv (`sim.base_models.Node`+): Node(s) to establish connections with.
+        """
+        for node in argv:
+            self.outs[node.id] = node
+            node.ins[self.id] = self
